fixing new_vis_modal
