+++ src/legacy/core_plugins/kibana/public/visualize/wizard/show_new_vis.tsx
@@ -21,15 +21,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { I18nContext } from 'ui/i18n';
-import { VisType } from 'ui/vis';
 import { NewVisModal } from './new_vis_modal';
+import { TypesStart } from '../../../../visualizations/public/np_ready/types';
 
 interface ShowNewVisModalParams {
   editorParams?: string[];
 }
 
 export function showNewVisModal(
-  visTypeRegistry: VisType[],
+  visTypeRegistry: TypesStart,
   { editorParams = [] }: ShowNewVisModalParams = {}
 ) {
   const container = document.createElement('div');
